requirements.txt
mlflow>=3.11.0
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.8.0
pandas>=2.2.0
numpy>=1.26.0
